모듈화 및 디자인 수정

+++ service/qr/services.ts
@@ -1,19 +1,15 @@
 import Service from "../service";
 
-type InsertQRProb = {
+export type InsertQRProb = {
   id: string;
   qrName: string;
   original_id: string;
   user_id: string;
 };
 
-type UpdateQRProb = {
+export type UpdateQRProb = {
   newName: string;
   newUrl: string;
-  selectedQRId: string;
-};
-
-type DeleteQRProb = {
   selectedQRId: string;
 };
 
@@ -27,8 +23,7 @@ interface QRCodeData {
 
 class QRService extends Service {
   async insertQR(insertData: InsertQRProb) {
-    const supabase = await this.supabase;
-    const { error } = await supabase.from("qr_codes").insert([
+    const { error } = await this.supabase.from("qr_codes").insert([
       {
         id: insertData.id,
         name: insertData.qrName,
@@ -37,26 +32,23 @@ class QRService extends Service {
       },
     ]);
 
-    if (!error) {
+    if (error) {
+      throw error;
+    } else {
       const qrUrl = `ald.my/${insertData.id}`;
       alert("QR 코드가 저장되었습니다.");
       return qrUrl;
-    } else {
-      console.error(error);
-      return null;
     }
   }
 
   async fetchQR(user_id: string) {
-    const supabase = await this.supabase;
-    const { data, error } = await supabase
+    const { data, error } = await this.supabase
       .from("qr_codes")
       .select("*")
       .eq("user_id", user_id);
 
     if (error) {
-      console.error("Error fetching QR codes:", error.message);
-      return null;
+      throw error;
     } else {
       const qrCodesWithUrl = data.map((qrCode: QRCodeData) => ({
         ...qrCode,
@@ -67,30 +59,30 @@ class QRService extends Service {
   }
 
   async updateQR(updateData: UpdateQRProb) {
-    const supabase = await this.supabase;
-    const { error } = await supabase
+    const { error } = await this.supabase
       .from("qr_codes")
       .update({ name: updateData.newName, original_id: updateData.newUrl })
       .eq("id", updateData.selectedQRId);
 
     if (error) {
-      console.error("Error updating QR code:", error.message);
-      return error;
+      throw error;
     } else {
       return null;
     }
   }
 
-  async deleteQR(deleteData: DeleteQRProb) {
-    const supabase = await this.supabase;
-    const { error } = await supabase
+  async deleteQR(userId: string, selectedQRId: string) {
+    console.log('userId : ', userId)
+    console.log('selectedQRId : ', selectedQRId)
+    const { error } = await this.supabase
       .from("qr_codes")
       .delete()
-      .eq("id", deleteData.selectedQRId)
+      .eq("id", selectedQRId)
+      .eq("user_id", userId);
 
     if (error) {
       console.error("Error deleting QR code:", error.message);
-      return error;
+      throw error;
     } else {
       return null;
     }
